Add Buttons and Change some Fonts

+++ src/Project/Utils.java
@@ -19,11 +19,11 @@ public class Utils {
         return (int) ((y + r) / 10) - 1;
     }
 
-    public static double arcTrX(double i) {
+    public static double arcTrX(int i) {
         return (i + 1) * 10.0 - r;
     }
 
-    public static double arcTrY(double j) {
+    public static double arcTrY(int j) {
         return (j + 1) * 10.0 - r;
     }
